#325 [feat] : TimeFormat mapping 함수 분리
